+++ setup.cfg
@@ -49,7 +49,7 @@ where = src
 
 [options.entry_points]
 console_scripts =
-        dakara-feed = dakara_feeder.commands.feed:main
+        dakara-feeder = dakara_feeder.__main__:main
 
 [flake8]
 max-line-length = 88
