--- src/dakara_feeder/commands/__init__.py
@@ -1,11 +1,0 @@
-from dakara_feeder.commands.create_config import CreateConfigSubcommand
-from dakara_feeder.commands.songs import SongsSubcommand
-from dakara_feeder.commands.tags import TagsSubcommand
-from dakara_feeder.commands.work_types import WorkTypesSubcommand
-
-__all__ = [
-    "CreateConfigSubcommand",
-    "SongsSubcommand",
-    "TagsSubcommand",
-    "WorkTypesSubcommand",
-]
